upload(or)download/requirements.txt
flask==2.3.3
python-dotenv==1.0.1
mysql-connector-python==8.2.0
werkzeug==2.3.8
pygithub==2.1.1
